backend/requirements.txt
asgiref==3.8.1
backports.statistics==0.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
Cython==3.0.10
Django>=3.2.25
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt
dl==0.1.0
docopt==0.6.2
gunicorn>=22.0.0
idna==3.7
importlib-metadata==7.1.0
ipaddr==2.2.0
ordereddict==1.1
packaging==24.0
pipreqs>=0.4.12
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3>=1.26.18
whitenoise==6.6.0
yarg==0.1.9
zipp==3.18.1
